Move EvolvedOpAnsatz to circuit.library.ansaetze
